Use Square enum and add left and right shift operator

+++ src/main.cpp
@@ -3,10 +3,10 @@
 int main() {
     chess_engine::Bitboard bitboard;
     bitboard.print();
-    bitboard.setBit(static_cast<int>(chess_engine::Square::e4));
+    bitboard.setBit(chess_engine::Square::e4);
     bitboard.print();
-    bitboard.setBit(static_cast<int>(chess_engine::Square::d5));
+    bitboard.setBit(chess_engine::Square::d5);
     bitboard.print();
-    bitboard.clearBit(static_cast<int>(chess_engine::Square::e4));
+    bitboard.clearBit(chess_engine::Square::e4);
     bitboard.print();
 }
